Рефакторинг для тестов.

+++ internal/storage/migrations/20250421234705_create_orders_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS orders (
     number        BIGINT                   NOT NULL PRIMARY KEY,
     user_id       UUID                     NOT NULL,
     status        TEXT                     NOT NULL,
-    accrual       INT                      NULL,
+    accrual       DECIMAL(12, 2)               NULL,
     uploaded_at   TIMESTAMP WITH TIME ZONE NOT NULL
 );
 -- +goose StatementEnd
